+++ kwdagger/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-__version__ = '0.2.7'
+__version__ = '0.3.0'
 
 __autogen__ = """
 mkinit  ~/code/kwdagger/kwdagger/__init__.py -w